+++ index.js
@@ -2,14 +2,23 @@ const Ofn = require( 'oro-functions' );
 const Sentry = require( '@sentry/node' );
 const { ExtraErrorData } = require( '@sentry/integrations' );
 
+const DEFAULT_TAGS = [ 'projectname', 'projectserver', 'lang', 'database', 'action', 'task' ];
+
 class OSentry {
 
-    static LEVEL = { DEBUG: 'debug', INFO: 'info', WARNING: 'warning', ERROR: 'error', FATAL: 'fatal' };
+    static LEVEL = {
+        LOG: 'log',
+        INFO: 'info',
+        DEBUG: 'debug',
+        WARNING: 'warning',
+        ERROR: 'error',
+        FATAL: 'fatal'
+    };
 
     #dsn;
     #status;
     #environment;
-    #constOptions;
+    #defaultOptions;
     #sentryOptions;
 
     defaultTags;
@@ -25,10 +34,9 @@ class OSentry {
         this.#environment = environment || 'UNDEFINED';
         this.projectname = projectname;
         this.projectserver = projectserver;
-        this.defaultTags = Ofn.isArray( defaultTags ) ? defaultTags :
-            [ 'projectname', 'projectserver', 'lang', 'database', 'action', 'task' ];
+        this.defaultTags = Ofn.cloneObject( Ofn.isArray( defaultTags ) ? defaultTags : DEFAULT_TAGS );
 
-        this.#constOptions = options;
+        this.#defaultOptions = options;
 
         if( this.#dsn && autoInit ) {
             this.init();
@@ -46,34 +54,40 @@ class OSentry {
     init( options = {} ) {
         if( this.#status ) { return Ofn.setResponseKO( 'OSentry is already init.' ) }
 
-        let cloneOpts = Object.assign(
-            {}, this.#constOptions, Ofn.isObject( options ) ? options : {} );
+        const defaultOptions = Ofn.cloneObject( {
+            ...this.#defaultOptions,
+            ...(Ofn.isObject( options ) ? options : {})
+        } );
 
-        ! this.#dsn && Ofn.isString( cloneOpts.dsn ) && ( this.#dsn = cloneOpts.dsn );
-        cloneOpts.dsn && ( delete cloneOpts.dsn );
+        ! this.#dsn && Ofn.isString( defaultOptions.dsn ) && ( this.#dsn = defaultOptions.dsn );
+        defaultOptions.dsn && ( delete defaultOptions.dsn );
 
-        Ofn.isString( cloneOpts.environment ) && ( this.#environment = cloneOpts.environment );
-        cloneOpts.environment && ( delete cloneOpts.environment );
+        Ofn.isString( defaultOptions.environment ) && ( this.#environment = defaultOptions.environment );
+        defaultOptions.environment && ( delete defaultOptions.environment );
 
         if( ! this.#dsn ) {
             console.error( 'OSentry: Sentry need DSN.' );
             return Ofn.setResponseKO( 'OSentry: Sentry need DSN.' );
         }
 
-        ! Ofn.isUndefined( cloneOpts.normalizeDepth ) && ! Ofn.isNumeric( cloneOpts.normalizeDepth ) && ( cloneOpts.normalizeDepth = 11 );
-        let normalizeDepth = cloneOpts.normalizeDepth || 11;
+        ! Ofn.isUndefined( defaultOptions.normalizeDepth ) && ! Ofn.isNumeric( defaultOptions.normalizeDepth ) && ( defaultOptions.normalizeDepth = 11 );
+        const normalizeDepth = defaultOptions.normalizeDepth || 11;
+        delete defaultOptions.normalizeDepth;
 
-        this.#sentryOptions = Object.assign( {
+        this.#sentryOptions = {
             dsn: this.#dsn,
             tracesSampleRate: 0.5,
             environment: this.#environment,
             integrations: [ new ExtraErrorData( { depth: normalizeDepth - 1 } ) ],
-            normalizeDepth
-        }, cloneOpts );
+            normalizeDepth,
+            ...defaultOptions,
+        };
 
-        try { Sentry.init( this.#sentryOptions ); }
+        try {
+            Sentry.init( this.#sentryOptions );
+        }
         catch( err ) {
-            let msg = err.toString().split( '\r\n' )[ 0 ].replace( '\n', ' ' );
+            const msg = err.toString().split( '\r\n' )[ 0 ].replace( '\n', ' ' );
             return Ofn.setResponseKO( `OSentry init failed: ${msg}`, { sentry: err } );
         }
 
@@ -96,10 +110,10 @@ class OSentry {
             Object.assign( extra, Ofn.cloneObject( response.error ), extra );
         }
 
-        let msg = extra.msg || extra.message || 'OSentry: response without msg';
+        const msg = extra.msg || extra.message || 'OSentry: response without msg';
         delete extra.msg;
 
-        let level = extra.level || ( extra.status ? OSentry.LEVEL.INFO : OSentry.LEVEL.ERROR );
+        const level = extra.level || ( extra.status ? OSentry.LEVEL.INFO : OSentry.LEVEL.ERROR );
         delete extra.level;
         delete extra.status;
 
@@ -109,7 +123,7 @@ class OSentry {
         let user = Ofn.cloneObjectWithKeys( extra.user, [ 'id', 'username', 'email' ] );
         Ofn.objIsEmpty( extra.user ) && ( delete extra.user );
 
-        let tags = Ofn.cloneObject( extra.tags || {} );
+        const tags = Ofn.cloneObject( extra.tags || {} );
         delete extra.tags;
 
         this.defaultTags.forEach( tagKey => {
@@ -131,7 +145,7 @@ class OSentry {
             return Ofn.setResponseKO( 'OSentry: not init (captureMessage).' );
         }
 
-        let id = Sentry.captureMessage( message, captureContext );
+        const id = Sentry.captureMessage( message, captureContext );
 
         return Ofn.setResponseOK( { event: { id, message, captureContext } } );
     }